refactor: merge parcel operators fetcher and parcel permissions fetcher
