+++ src/components.ts
@@ -38,8 +38,7 @@ import { IFetchComponent } from '@well-known-components/interfaces'
 import { createAlchemyNftFetcher } from './adapters/alchemy-nft-fetcher'
 import { createThirdPartyContractRegistry } from './ports/ownership-checker/third-party-contract-registry'
 import { createThirdPartyItemChecker } from './ports/ownership-checker/third-party-item-checker'
-import { createParcelPermissionsComponent } from './adapters/parcel-permissions-fetcher'
-import { createParcelOperatorsComponent } from './adapters/parcel-operators-fetcher'
+import { createParcelRightsComponent } from './adapters/parcel-rights-fetcher'
 
 // Initialize all the components of the app
 export async function initComponents(
@@ -96,8 +95,7 @@ export async function initComponents(
   )
   const namesFetcher = createElementsFetcherComponent({ logs }, async (address) => fetchAllNames({ theGraph }, address))
   const landsFetcher = createElementsFetcherComponent({ logs }, async (address) => fetchAllLANDs({ theGraph }, address))
-  const parcelPermissionsFetcher = await createParcelPermissionsComponent({ logs, theGraph })
-  const parcelOperatorsFetcher = await createParcelOperatorsComponent({ logs, theGraph })
+  const parcelRightsFetcher = await createParcelRightsComponent({ logs, theGraph })
 
   const resourcesStatusCheck = createResourcesStatusComponent({ logs })
   const status = await createStatusComponent({ logs, fetch })
@@ -183,8 +181,7 @@ export async function initComponents(
     emotesFetcher,
     namesFetcher,
     landsFetcher,
-    parcelPermissionsFetcher,
-    parcelOperatorsFetcher,
+    parcelRightsFetcher,
     thirdPartyProvidersGraphFetcher,
     thirdPartyProvidersStorage,
     contentServerUrl,
