+++ src/types.ts
@@ -36,8 +36,7 @@ import { ThirdPartyProvidersStorage } from './logic/third-party-providers-storag
 import { IProfilesComponent } from './adapters/profiles'
 import { AlchemyNftFetcher } from './adapters/alchemy-nft-fetcher'
 import { ThirdPartyItemChecker } from './ports/ownership-checker/third-party-item-checker'
-import { ParcelPermissionsFetcher } from './adapters/parcel-permissions-fetcher'
-import { ParcelOperatorsFetcher } from './adapters/parcel-operators-fetcher'
+import { ParcelRightsFetcher } from './adapters/parcel-rights-fetcher'
 
 export type GlobalContext = {
   components: BaseComponents
@@ -65,8 +64,7 @@ export type BaseComponents = {
   entitiesFetcher: EntitiesFetcher
   namesFetcher: ElementsFetcher<Name>
   landsFetcher: ElementsFetcher<LAND>
-  parcelPermissionsFetcher: ParcelPermissionsFetcher
-  parcelOperatorsFetcher: ParcelOperatorsFetcher
+  parcelRightsFetcher: ParcelRightsFetcher
   resourcesStatusCheck: IResourcesStatusComponent
   status: IStatusComponent
   l1Provider: HTTPProvider
